Reset Added and Difficulty Reduced
Difficulty rating was too strict, so that's been simplified.
The ability to reset the puzzle at any time has been added; this does not restart the timer.
The main menu is now accessible from the pause/help screen to avoid the need to either complete the puzzle or quit and reopen the program. 
Other small wording changes to make the UI more clear.