Colours
User can now switch to different colours for inputs. This can be useful if they are unsure of some numbers, they can add them in another colour so that if it was a mistake, they can go back and change the appropriate ones. Also useful if the background colour of the user's terminal is not suitable for the default cyan input colour.